VideoPlayer: move FrameBufferOject to VideoPlayer, its the only client
